+++ src/Resistor.java
@@ -9,8 +9,8 @@ public class Resistor extends Element {
         node1 = info[1];
         node2 = info[2];
         if (info[3].charAt(info[3].length() - 1) >= '0' && info[3].charAt(info[3].length() - 1) <= '9')
-            resistance = Integer.parseInt(info[3]);
-        else resistance = Integer.parseInt(info[3].substring(0, info[3].length() - 1));
+            resistance = Double.parseDouble(info[3]);
+        else resistance = Double.parseDouble(info[3].substring(0, info[3].length() - 1));
         switch (info[3].charAt(info[3].length() - 1)) {
             case 'p':
                 resistance *= 1e-12;
